complete scheduled (need add method for second update state of report need fix writing dates to table)

+++ src/main/java/com/bot/employeeTimeTracongBot/model/User.java
@@ -15,7 +15,7 @@ public class User {
     private boolean isSendReport;
     private String name;
     private String dateConnecting;
-    private int chatId;
+    private long chatId;
     private String nickName;
     private String fullName;
     private String dateLastReport;
